MpaDecFilter: updated "#include ..." DeCSS: fixed warning.

+++ src/DeCSS/DeCSSInputPin.cpp
@@ -170,7 +170,7 @@ void CDeCSSInputPin::StripPacket(BYTE*& p, long& len)
             }
 
             if (expected > 0) {
-                expected -= (p - p0);
+                expected -= (int)(p - p0);
                 len = min(expected, len);
             }
         }
